添加日志监控

+++ manager/ar.py
@@ -6,23 +6,36 @@
 
 import os, traceback
 from login.models import *
-from manager.models import Role,User_UIControl
+from manager.models import *
+from login.models import *
 from django.db.models import Q
-from manager.context import Me2Log as logger,get_temp_dir
+from manager.context import Me2Log as logger,get_temp_dir,monitor
 
 class RoleData():
     '''
     角色操作
     '''
-
     @classmethod
+    @monitor(action='查询角色详情')
     def queryroletable(cls,**kws):
+        searchvalue=kws.get('searchvalue','')
+        data=[]
+        data0=Role.objects.filter(Q(name__icontains=searchvalue)|Q(description__icontains=searchvalue))
+        for x in data0:
+            data.append({
+                'id':x.id,
+                'name':x.name,
+                'description':x.description,
+                'createtime':x.createtime,
+                'authorname':x.author.name
+
+                })
         try:
             return{
                 'code':0,
                 'msg':'获得权限表',
-
-                'data':list(Role.objects.all())
+                'data':data,
+                'count':len(data)
             }
         except:
             return{
@@ -33,17 +46,20 @@ class RoleData():
     
 
     @classmethod
+    @monitor(action='添加角色')
     def addrole(cls,**kws):
         try:
             role=Role()
+            logger.error(kws)
             role.name=kws['name']
             role.description=kws['description']
             role.author=kws['user']
             role.save()
             userids=kws['userids'].split(',')
-            if(userids!=[]):
+            if(userids):
                 for idk in userids:
-                    role.users.add(User.objects.get(id=idk))
+                    if idk:
+                        role.users.add(User.objects.get(id=idk))
 
 
             return {
@@ -59,22 +75,26 @@ class RoleData():
 
 
     @classmethod
+    @monitor(action='删除角色')
     def delrole(cls,**kws):
         try:
             role_ids=kws['ids']
             for roleid in role_ids.split(','):
-                Role.objects.get(id=roleid).delete()
+                if roleid:
+                    Role.objects.get(id=roleid).delete()
             return{
                 'code':0,
                 'msg':'删除角色成功'
             }
         except:
+            logger.error('删除角色异常：',traceback.format_exc())
             return{
              'code':4,
              'msg':'删除角色异常'
             }
 
     @classmethod
+    @monitor(action='更新角色',description='$Role.objects.get(id=[uid])')
     def updaterole(cls,**kws):
         try:
             role_id=kws['uid']
@@ -100,6 +120,7 @@ class RoleData():
             }
 
     @classmethod
+    @monitor(action='查询角色详情')
     def queryonerole(cls,**kws):
         try:
             role_id=kws['uid']
@@ -137,9 +158,7 @@ class Grant(object):
     ''' 
     @classmethod
     def isconfig(cls,code):
-
         scandirs=get_temp_dir()
-        logger.info('UI组件扫描位置:',scandirs)
         for path in scandirs:
             for filename in os.listdir(path):
                 if filename.split('.')[1]=='html':
@@ -159,7 +178,7 @@ class Grant(object):
         if uc.exists():
             isopen=uc[0].is_open
         else:
-            logger.info('组件[%s]没配置 放行'%code)
+            #logger.info('组件[%s]没配置 放行'%code)
             return '!important'
 
         user=User.objects.get(name=username)
@@ -170,19 +189,19 @@ class Grant(object):
             if user in r.users.all():
                 user_role_ids.append(r.id)
 
-        logger.info('用户[%s]ID[%s]'%(username,User.objects.get(name=username).id))
-        logger.info('用户[%s]角色ID:%s'%(username,user_role_ids))
+        # logger.info('用户[%s]ID[%s]'%(username,User.objects.get(name=username).id))
+        # logger.info('用户[%s]角色ID:%s'%(username,user_role_ids))
         f1=User_UIControl.objects.filter(user_id=user_id,kind='USER')
         if f1.exists() and isopen:
-            logger.info('用户[%s]看不到UI组件[%s]'%(username,uc[0].description))
+            #logger.info('用户[%s]看不到UI组件[%s]'%(username,uc[0].description))
             return 'none!important'
 
         for idx in user_role_ids:
             f2=User_UIControl.objects.filter(user_id=idx,kind='ROLE')
             if f2.exists() and isopen:
-                logger.info('角色[%s]看不到UI组件[%s]'%(Role.objects.get(id=idx).name,uc[0].description))
+                #logger.info('角色[%s]看不到UI组件[%s]'%(Role.objects.get(id=idx).name,uc[0].description))
                 return 'none!important'
-        logger.info('用户[%s]能看到UI组件[%s]'%(username,uc[0].description))
+        #logger.info('用户[%s]能看到UI组件[%s]'%(username,uc[0].description))
         return '!important'
     
     @classmethod
@@ -224,6 +243,7 @@ class Grant(object):
             raise RuntimeError('更新视图控制用户信息异常')
     
     @classmethod
+    @monitor(action='查询权限详情')
     def query_one_ui_control(cls,uid):
         try:
             u=UIControl.objects.get(id=uid)
@@ -259,6 +279,7 @@ class Grant(object):
 
 
     @classmethod
+    @monitor(action='添加权限')
     def add_ui_control(cls, **config):
         try:
             uc = UIControl()
@@ -282,6 +303,7 @@ class Grant(object):
             }
     
     @classmethod
+    @monitor(action='删除权限')
     def del_ui_control(cls, **kws):
         try:
             ucid=kws['ucids']
@@ -307,6 +329,7 @@ class Grant(object):
             }
     
     @classmethod
+    @monitor(action='编辑权限')
     def edit_ui_control(cls, **config):
         try:
             uc = UIControl.objects.get(id=config['cid'])
@@ -330,7 +353,10 @@ class Grant(object):
             }
     
     @classmethod
-    def query_ui_grant_table(cls, searchvalue=None):
+    @monitor(action='查询权限列表')
+    def query_ui_grant_table(cls, **kws):
+
+        searchvalue=kws.get('searchvalue','')
         try:
             data = list()
             uicall = []
@@ -352,7 +378,8 @@ class Grant(object):
             return {
                 'code': 0,
                 'msg': '',
-                'data': data
+                'data': data,
+                'count':len(data)
             }
         except:
             logger.error('获取权限表查询异常:',traceback.format_exc())
@@ -388,6 +415,7 @@ class Grant(object):
         }
 
     @classmethod
+    @monitor(action='开关权限')
     def updateuicontrolstatus(cls,**kws):
         try:
             u=UIControl.objects.get(id=kws['uid'])
